converting all the lwrp files to custom resources. helpers still need work.
