--- libraries/pacifica_reporting.rb
@@ -1,12 +1,0 @@
-# pacifica cookbook module
-module PacificaCookbook
-  require_relative 'pacifica_base_php'
-  # installs and configures pacifica cart frontend wsgi
-  class PacificaReporting < PacificaBasePhp
-    property :name, String, name_property: true
-    property :git_opts, Hash, default: {
-      repository: 'https://github.com/EMSL-MSC/pacifica-reporting.git'
-    }
-    resource_name :pacifica_reporting
-  end
-end
